Flatten structure
Related to #13
