+++ templates_test.go
@@ -31,7 +31,7 @@
 // The fact that you are presently reading this means that you have had
 // knowledge of the CeCILL license and that you accept its terms.
 
-package templates
+package main
 
 import (
 	"bytes"
