+++ types.go
@@ -31,16 +31,82 @@
 // The fact that you are presently reading this means that you have had
 // knowledge of the CeCILL license and that you accept its terms.
 
-package types
+// Data structure representaton of the configuration
+
+package main
 
 import (
 	"fmt"
-	"go.nc0.fr/svgu/pkg/templates"
 	"os"
 	"path"
 	"strings"
 	"sync"
 )
+
+// Index is the global registry of Go modules.
+type Index struct {
+	Domain  string
+	Modules map[string]*Module
+	// internal
+	lock sync.Mutex
+}
+
+// AddModule adds a module to the index.
+func (i *Index) AddModule(n string, m *Module) {
+	i.lock.Lock()
+	defer i.lock.Unlock()
+	i.Modules[n] = m
+}
+
+// GetModule returns a module from the index.
+func (i *Index) GetModule(n string) *Module {
+	i.lock.Lock()
+	defer i.lock.Unlock()
+	return i.Modules[n]
+}
+
+// RemoveModule removes a module from the index.
+func (i *Index) RemoveModule(n string) {
+	i.lock.Lock()
+	defer i.lock.Unlock()
+	delete(i.Modules, n)
+}
+
+// CheckModule checks if a module is in the index.
+func (i *Index) CheckModule(n string) bool {
+	i.lock.Lock()
+	defer i.lock.Unlock()
+	_, ok := i.Modules[n]
+	return ok
+}
+
+// GenerateFile generates the index file.
+func (i *Index) GenerateFile(out string) error {
+	i.lock.Lock()
+	defer i.lock.Unlock()
+
+	f := path.Join(out, "index.html")
+
+	// Create the file.
+	fd, err := os.Create(f)
+	if err != nil {
+		return err
+	}
+	defer func(fd *os.File) {
+		err := fd.Close()
+		if err != nil {
+			panic(err)
+		}
+	}(fd)
+
+	// Execute the template and write the output to the file.
+	if err := ExecIndex(fd,
+		"https://pkg.go.dev", 2); err != nil {
+		return err
+	}
+
+	return nil
+}
 
 // Vcs is an enum for version control systems supported by the standard Go
 // toolchain.
@@ -100,7 +166,7 @@ func (m *Module) GenerateFile(out string, domain string) error {
 	}(fd)
 
 	// Execute the template and write the output to the file.
-	if err := templates.ExecModule(fd,
+	if err := ExecModule(fd,
 		fmt.Sprintf("%s/%s", domain, p), string(v), r,
 		d, f); err != nil {
 		return err
